+++ zmq/backend/cffi/_cffi.py
@@ -49,6 +49,7 @@ int zmq_socket_monitor(void *socket, const char *addr, int events);
 core40_functions = \
 '''
 int zmq_curve_keypair (char *z85_public_key, char *z85_secret_key);
+int zmq_has (const char *capability);
 '''
 
 message32_functions = \